refactored user settings mergerd profile and user setttings
